Add solution button option and refactor
