+++ AngularApp/projects/diagnostic-data/src/lib/models/time-series.ts
@@ -101,3 +101,11 @@ export interface TablePoint {
   column: string;
   counterName: string;
 }
+
+export enum MetricType {
+  None,
+  Avg,
+  Min,
+  Max,
+  Sum,
+}
